+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/settings/SocialLinkViewModel.kt
@@ -18,7 +18,6 @@ import com.kakao.sdk.user.UserApiClient
 import com.wafflestudio.snutt2.data.user.UserRepository
 import com.wafflestudio.snutt2.lib.network.dto.GetUserFacebookResults
 import com.wafflestudio.snutt2.lib.network.dto.core.UserDto
-import com.wafflestudio.snutt2.ui.state.SocialLoginState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -30,25 +29,6 @@ class SocialLinkViewModel @Inject constructor(
     private val userRepository: UserRepository,
 ) : ViewModel() {
     val userInfo: StateFlow<UserDto?> = userRepository.user
-
-    val kakaolLoginState = MutableStateFlow<SocialLoginState>(SocialLoginState.Initial)
-    val googleLoginState = MutableStateFlow<SocialLoginState>(SocialLoginState.Initial)
-
-    private val loginWithKakaoAccountCallback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
-        if (error != null) {
-            if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
-                updateKakaoLoginState(SocialLoginState.Cancelled)
-            } else if (error is AuthError && error.reason == AuthErrorCause.AccessDenied) {
-                updateKakaoLoginState(SocialLoginState.Cancelled)
-            } else {
-                updateKakaoLoginState(SocialLoginState.Failed)
-            }
-        } else if (token != null) {
-            updateKakaoLoginState(SocialLoginState.Success(token.accessToken))
-        } else {
-            updateKakaoLoginState(SocialLoginState.Failed)
-        }
-    }
 
     suspend fun fetchUserInfo() {
         userRepository.fetchUserInfo()
@@ -64,63 +44,5 @@ class SocialLinkViewModel @Inject constructor(
 
     suspend fun disconnectFacebook() {
         userRepository.deleteUserFacebook()
-    }
-
-    fun updateKakaoLoginState(state: SocialLoginState) {
-        viewModelScope.launch {
-            kakaolLoginState.emit(state)
-        }
-    }
-
-    fun updateGoogleLoginState(state: SocialLoginState) {
-        viewModelScope.launch {
-            googleLoginState.emit(state)
-        }
-    }
-
-    fun triggerKakaoSignin(context: Context) {
-        updateKakaoLoginState(SocialLoginState.InProgress)
-        if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
-            UserApiClient.instance.loginWithKakaoTalk(context) { token, loginError ->
-                if (loginError != null) {
-                    if (loginError is ClientError && loginError.reason == ClientErrorCause.Cancelled) {
-                        updateKakaoLoginState(SocialLoginState.Cancelled)
-                    } else if (loginError is AuthError && loginError.reason == AuthErrorCause.AccessDenied) {
-                        updateKakaoLoginState(SocialLoginState.Cancelled)
-                    } else {
-                        // 카카오계정으로 로그인
-                        UserApiClient.instance.loginWithKakaoAccount(context = context, callback = loginWithKakaoAccountCallback)
-                    }
-                } else if (token != null) {
-                    updateKakaoLoginState(SocialLoginState.Success(token.accessToken))
-                } else {
-                    updateKakaoLoginState(SocialLoginState.Failed)
-                }
-            }
-        } else {
-            // 카카오계정으로 로그인
-            UserApiClient.instance.loginWithKakaoAccount(context = context, callback = loginWithKakaoAccountCallback)
-        }
-    }
-
-    fun handleGoogleLoginActivityResult(result: ActivityResult) {
-        if (result.resultCode == Activity.RESULT_OK) {
-            val task: Task<GoogleSignInAccount> = GoogleSignIn.getSignedInAccountFromIntent(result.data)
-            try {
-                val account = task.getResult(ApiException::class.java)
-                val authCode = account?.serverAuthCode
-                if (authCode != null) {
-                    updateGoogleLoginState(SocialLoginState.Success(authCode))
-                } else {
-                    updateGoogleLoginState(SocialLoginState.Failed)
-                }
-            } catch (e: ApiException) {
-                updateGoogleLoginState(SocialLoginState.Failed)
-            }
-        } else if (result.resultCode == Activity.RESULT_CANCELED) {
-            updateGoogleLoginState(SocialLoginState.Cancelled)
-        } else {
-            updateGoogleLoginState(SocialLoginState.Failed)
-        }
     }
 }
