+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_out/TutorialPage.kt
@@ -7,8 +7,6 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -19,10 +17,18 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.common.api.ApiException
+import com.google.android.gms.tasks.Task
+import com.kakao.sdk.auth.model.OAuthToken
+import com.kakao.sdk.common.model.AuthError
+import com.kakao.sdk.common.model.AuthErrorCause
+import com.kakao.sdk.common.model.ClientError
+import com.kakao.sdk.common.model.ClientErrorCause
+import com.kakao.sdk.user.UserApiClient
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.components.compose.BorderButton
 import com.wafflestudio.snutt2.components.compose.DividerWithText
@@ -32,10 +38,8 @@ import com.wafflestudio.snutt2.lib.android.toast
 import com.wafflestudio.snutt2.lib.facebookLogin
 import com.wafflestudio.snutt2.ui.SNUTTColors
 import com.wafflestudio.snutt2.ui.SNUTTTypography
-import com.wafflestudio.snutt2.ui.state.SocialLoginState
 import com.wafflestudio.snutt2.views.*
 import com.wafflestudio.snutt2.views.logged_in.home.HomeViewModel
-import com.wafflestudio.snutt2.views.logged_in.home.settings.SocialLinkViewModel
 import com.wafflestudio.snutt2.views.logged_in.home.settings.UserViewModel
 import kotlinx.coroutines.launch
 
@@ -50,10 +54,6 @@ fun TutorialPage() {
 
     val userViewModel = hiltViewModel<UserViewModel>()
     val homeViewModel = hiltViewModel<HomeViewModel>()
-    val socialLinkViewModel = hiltViewModel<SocialLinkViewModel>()
-
-    val kakaoLoginState by socialLinkViewModel.kakaolLoginState.collectAsStateWithLifecycle()
-    val googleLoginState by socialLinkViewModel.googleLoginState.collectAsStateWithLifecycle()
 
     val clientId = context.getString(R.string.web_client_id)
     val clientSecret = context.getString(R.string.web_client_secret)
@@ -75,7 +75,7 @@ fun TutorialPage() {
         }
     }
 
-    val loginWithGoogleAuthCode: (String) -> Unit = { authCode: String ->
+    val handleGoogleSignInCallback: (String) -> Unit = { authCode: String ->
         coroutineScope.launch {
             launchSuspendApi(
                 apiOnProgress = apiOnProgress,
@@ -101,7 +101,24 @@ fun TutorialPage() {
     val googleLoginActivityResultLauncher = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.StartActivityForResult(),
     ) { result ->
-        socialLinkViewModel.handleGoogleLoginActivityResult(result)
+        if (result.resultCode == Activity.RESULT_OK) {
+            val task: Task<GoogleSignInAccount> = GoogleSignIn.getSignedInAccountFromIntent(result.data)
+            try {
+                val account = task.getResult(ApiException::class.java)
+                val authCode = account?.serverAuthCode
+                if (authCode != null) {
+                    handleGoogleSignInCallback(authCode)
+                } else {
+                    context.toast(context.getString(R.string.sign_in_sign_in_google_failed_unknown))
+                }
+            } catch (e: ApiException) {
+                context.toast(context.getString(R.string.sign_in_sign_in_google_failed_unknown))
+            }
+        } else if (result.resultCode == Activity.RESULT_CANCELED) {
+            context.toast(context.getString(R.string.sign_in_sign_in_google_cancelled))
+        } else {
+            context.toast(context.getString(R.string.sign_in_sign_in_google_failed_unknown))
+        }
     }
 
     val googleSignInClient: GoogleSignInClient = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
@@ -112,9 +129,8 @@ fun TutorialPage() {
         }
 
     // 구글 계정 선택 창 띄움
-    val startGoogleSignIn = {
+    val handleGoogleSignIn = {
         val signInIntent = googleSignInClient.signInIntent
-        socialLinkViewModel.updateGoogleLoginState(SocialLoginState.InProgress)
         googleLoginActivityResultLauncher.launch(signInIntent)
     }
 
@@ -129,47 +145,50 @@ fun TutorialPage() {
                     userViewModel.loginKakao(kakaoAccessToken)
                     homeViewModel.refreshData()
                     navController.navigateAsOrigin(NavigationDestination.Home)
-                    socialLinkViewModel.updateKakaoLoginState(SocialLoginState.Initial)
                 } else {
-                    socialLinkViewModel.updateKakaoLoginState(SocialLoginState.Failed)
+                    context.toast(context.getString(R.string.sign_in_kakao_failed_unknown))
                 }
             }
         }
     }
 
-    LaunchedEffect(kakaoLoginState) {
-        when (kakaoLoginState) {
-            is SocialLoginState.Initial -> {}
-            is SocialLoginState.InProgress -> {}
-            is SocialLoginState.Cancelled -> {
+    val loginWithKakaoAccountCallback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
+        if (error != null) {
+            if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
                 context.toast(context.getString(R.string.sign_in_kakao_failed_cancelled))
-                socialLinkViewModel.updateKakaoLoginState(SocialLoginState.Initial)
-            }
-            is SocialLoginState.Failed -> {
+            } else if (error is AuthError && error.reason == AuthErrorCause.AccessDenied) {
+                context.toast(context.getString(R.string.sign_in_kakao_failed_cancelled))
+            } else {
                 context.toast(context.getString(R.string.sign_in_kakao_failed_unknown))
-                socialLinkViewModel.updateKakaoLoginState(SocialLoginState.Initial)
             }
-            is SocialLoginState.Success -> {
-                loginWithKaKaoAccessToken((kakaoLoginState as SocialLoginState.Success).token)
-            }
+        } else if (token != null) {
+            loginWithKaKaoAccessToken(token.accessToken)
+        } else {
+            context.toast(context.getString(R.string.sign_in_kakao_failed_unknown))
         }
     }
 
-    LaunchedEffect(googleLoginState) {
-        when (googleLoginState) {
-            is SocialLoginState.Initial -> {}
-            is SocialLoginState.InProgress -> {}
-            is SocialLoginState.Cancelled -> {
-                context.toast(context.getString(R.string.sign_in_sign_in_google_cancelled))
-                socialLinkViewModel.updateGoogleLoginState(SocialLoginState.Initial)
+    val handleKakaoSignin: () -> Unit = {
+        if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
+            UserApiClient.instance.loginWithKakaoTalk(context) { token, loginError ->
+                if (loginError != null) {
+                    if (loginError is ClientError && loginError.reason == ClientErrorCause.Cancelled) {
+                        context.toast(context.getString(R.string.sign_in_kakao_failed_cancelled))
+                    } else if (loginError is AuthError && loginError.reason == AuthErrorCause.AccessDenied) {
+                        context.toast(context.getString(R.string.sign_in_kakao_failed_cancelled))
+                    } else {
+                        // 카카오계정으로 로그인
+                        UserApiClient.instance.loginWithKakaoAccount(context = context, callback = loginWithKakaoAccountCallback)
+                    }
+                } else if (token != null) {
+                    loginWithKaKaoAccessToken(token.accessToken)
+                } else {
+                    context.toast(context.getString(R.string.sign_in_kakao_failed_unknown))
+                }
             }
-            is SocialLoginState.Failed -> {
-                context.toast(context.getString(R.string.sign_in_sign_in_google_failed_unknown))
-                socialLinkViewModel.updateGoogleLoginState(SocialLoginState.Initial)
-            }
-            is SocialLoginState.Success -> {
-                loginWithGoogleAuthCode((googleLoginState as SocialLoginState.Success).token)
-            }
+        } else {
+            // 카카오계정으로 로그인
+            UserApiClient.instance.loginWithKakaoAccount(context = context, callback = loginWithKakaoAccountCallback)
         }
     }
 
@@ -240,14 +259,14 @@ fun TutorialPage() {
             ) {
                 SocialLoginButton(
                     painter = painterResource(id = R.drawable.kakao_login),
-                    onClick = { socialLinkViewModel.triggerKakaoSignin(context) },
+                    onClick = { handleKakaoSignin() },
                 )
 
                 SocialLoginButton(
                     painter = painterResource(id = R.drawable.google_login),
                     onClick = {
                         googleSignInClient.signOut().addOnCompleteListener {
-                            startGoogleSignIn()
+                            handleGoogleSignIn()
                         }
                     },
                 )
